+++ src/main/java/com/empbulletin/bootcampersbulletin/repository/MarksRepository.java
@@ -1,10 +1,11 @@
 package com.empbulletin.bootcampersbulletin.repository;
-
-import com.empbulletin.bootcampersbulletin.model.Employee;
 import com.empbulletin.bootcampersbulletin.model.Marks;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MarksRepository extends JpaRepository<Marks, Integer> {
+import java.util.List;
+
+public interface MarksRepository extends JpaRepository<Marks, Long> {
+
 
 
 }
